+++ example/lib/pages/card.dart
@@ -106,7 +106,7 @@ class _CardPageState extends State<CardPage> {
                 child: Row(
                   children: [
                     ShadImage.square(
-                      ShadAssets.bellRing,
+                      LucideIcons.bellRing,
                       size: 24,
                       color: theme.colorScheme.foreground,
                     ),
@@ -183,7 +183,7 @@ class _CardPageState extends State<CardPage> {
             text: const Text('Mark all as read'),
             icon: const Padding(
               padding: EdgeInsets.only(right: 8),
-              child: ShadImage.square(ShadAssets.check, size: 16),
+              child: ShadImage.square(LucideIcons.check, size: 16),
             ),
             onPressed: () {},
           ),