+++ lib/shadcn_ui.dart
@@ -31,9 +31,6 @@ export 'src/components/switch.dart';
 export 'src/components/toast.dart';
 export 'src/components/tooltip.dart';
 
-// Assets
-export 'src/assets.dart';
-
 // Raw Components
 export 'src/raw_components/portal.dart';
 export 'src/raw_components/same_width_column.dart';
@@ -92,3 +89,4 @@ export 'src/utils/extensions.dart';
 export 'package:flutter_animate/flutter_animate.dart';
 export 'package:boxy/boxy.dart';
 export 'package:boxy/flex.dart';
+export 'package:lucide_icons_flutter/lucide_icons.dart';